refactor insert, delete

+++ app/src/main/java/com/lynaysereyvath/remindme/domain/QuoteEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "tblQuote")
 class QuoteEntity(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Long = 0,
     @ColumnInfo(name = "author") val author: String,
     @ColumnInfo(name = "message") val message: String
 )
